plat/common: Define __MAX_IRQ for different architectures
In current IRQ framework, we only support 16 irq_handlers. This is
well for x86 legacy interrupts. But for Arm, it's not enough. In
this case, we defined __MAX_IRQ 1020 for Arm and keep __MAX_IRQ 16
for x86. This will waste some space, but we can keep O(1) efficency.

Change-Id: Ib499718d9faeb1db27307b5e1c7a7f2e4b8ec293
Jira: ENTOS-655
Signed-off-by: Wei Chen <wei.chen@arm.com>